+++ components/shared/index.js
@@ -1,9 +1,13 @@
 import ErrorMessage from './ErrorMessage';
 import Label from './Label';
 import FieldGroup from './FieldGroup';
+import icons from './icons';
+import masks from './masks';
 
 export {
   ErrorMessage,
   Label,
   FieldGroup,
+  icons,
+  masks,
 };